Move LIN frame related stuff in frame module
Previously it was split in the master and top level module.
